requirements.txt
GitPython>=3.1.44
tree_sitter_python>=0.23.0
tree-sitter>=0.23.2
tqdm>=4.67.1
huggingface_hub
datasets
